Extract StatisticsProtocol due to multiple consumers
