+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -5,15 +5,7 @@ protocol QuestionLoading {
     func requestNextQuestion()
 }
 
-protocol StatisticsReporting {
-    var totalAccuracy: Double { get }
-    var gamesCount: Int { get }
-    var bestGame: GameRecord { get }
-}
 
-protocol StatisticsStoring {
-    func store(correct count: Int, total amount: Int)
-}
 
 protocol ResultPresenting {
     func displayResults(
